refactor: clean and move tests
