--- tests/models/test_eof_wrapper.py
@@ -1,105 +1,0 @@
-import numpy as np
-import pandas as pd
-import xarray as xr
-import pytest
-
-from xeofs.models.eof import EOF
-from xeofs.xarray.eof import EOF as xrEOF
-
-
-@pytest.mark.parametrize('scaling', [0, 1, 2])
-def test_wrapper_solutions(scaling, sample_array):
-    # Solutions of numpy, pandas and xarray wrapper are the same
-    X = sample_array
-    df = pd.DataFrame(X)
-    da = xr.DataArray(X)
-    # Perform analysis with all three wrappers
-    numpy_model = EOF(X)
-    numpy_model.solve()
-
-    xarray_model = xrEOF(da, dim='dim_0')
-    xarray_model.solve()
-
-    # Explained variance
-    desired_expvar = numpy_model.explained_variance()
-    actual_xarray_expvar = xarray_model.explained_variance()
-    # Explained variance ratio
-    desired_expvar_ratio = numpy_model.explained_variance_ratio()
-    actual_xarray_expvar_ratio = xarray_model.explained_variance_ratio()
-    # PCs
-    desired_pcs = numpy_model.pcs(scaling=scaling)
-    actual_xarray_pcs = xarray_model.pcs(scaling=scaling).values
-    # EOFs
-    desired_eofs = numpy_model.eofs(scaling=scaling)
-    actual_xarray_eofs = xarray_model.eofs(scaling=scaling).values
-    # EOFs as correlation
-    desired_eofs_corr = numpy_model.eofs_as_correlation()
-    actual_xarray_eofs_corr = xarray_model.eofs_as_correlation()
-    # Reconstructed X
-    desired_Xrec = numpy_model.reconstruct_X()
-    actual_xarray_Xrec = xarray_model.reconstruct_X()
-    # Projection onto EOFs
-    desired_proj = numpy_model.project_onto_eofs(X, scaling=scaling)
-    actual_xarray_proj = xarray_model.project_onto_eofs(da, scaling=scaling)
-
-    np.testing.assert_allclose(actual_xarray_expvar, desired_expvar)
-    np.testing.assert_allclose(actual_xarray_expvar_ratio, desired_expvar_ratio)
-    np.testing.assert_allclose(actual_xarray_pcs, desired_pcs)
-    np.testing.assert_allclose(actual_xarray_eofs, desired_eofs)
-    np.testing.assert_allclose(actual_xarray_eofs_corr[0], desired_eofs_corr[0])
-    np.testing.assert_allclose(actual_xarray_eofs_corr[1], desired_eofs_corr[1])
-    np.testing.assert_allclose(actual_xarray_Xrec, desired_Xrec)
-    np.testing.assert_allclose(actual_xarray_proj, desired_proj)
-
-
-@pytest.mark.parametrize('scaling', [0, 1, 2])
-def test_wrapper_multivariate_solutions(scaling, sample_array):
-    # Solutions of numpy, pandas and xarray wrapper are the same
-    X = sample_array
-    X1 = X[:, :10]
-    X2 = X[:, 10:]
-    df1 = pd.DataFrame(X1)
-    df2 = pd.DataFrame(X2)
-    da1 = xr.DataArray(X1)
-    da2 = xr.DataArray(X2)
-    # Perform analysis with all three wrappers
-    numpy_model = EOF([X1, X2])
-    numpy_model.solve()
-
-    xarray_model = xrEOF([da1, da2], dim='dim_0')
-    xarray_model.solve()
-
-    # Explained variance
-    desired_expvar = numpy_model.explained_variance()
-    actual_xarray_expvar = xarray_model.explained_variance()
-    # Explained variance ratio
-    desired_expvar_ratio = numpy_model.explained_variance_ratio()
-    actual_xarray_expvar_ratio = xarray_model.explained_variance_ratio()
-    # PCs
-    desired_pcs = numpy_model.pcs(scaling=scaling)
-    actual_xarray_pcs = xarray_model.pcs(scaling=scaling)
-    # EOFs
-    desired_eofs = numpy_model.eofs(scaling=scaling)
-    actual_xarray_eofs = xarray_model.eofs(scaling=scaling)
-    # EOFs as correlation
-    desired_eofs_corr = numpy_model.eofs_as_correlation()
-    actual_xarray_eofs_corr = xarray_model.eofs_as_correlation()
-    # Reconstructed X
-    desired_Xrec = numpy_model.reconstruct_X()
-    actual_xarray_Xrec = xarray_model.reconstruct_X()
-    # Projection onto EOFs
-    desired_proj = numpy_model.project_onto_eofs([X1, X2], scaling=scaling)
-    actual_xarray_proj = xarray_model.project_onto_eofs([da1, da2], scaling=scaling)
-
-    np.testing.assert_allclose(actual_xarray_expvar, desired_expvar)
-    np.testing.assert_allclose(actual_xarray_expvar_ratio, desired_expvar_ratio)
-    np.testing.assert_allclose(actual_xarray_pcs, desired_pcs)
-    np.testing.assert_allclose(actual_xarray_eofs[0], desired_eofs[0])
-    np.testing.assert_allclose(actual_xarray_eofs[1], desired_eofs[1])
-    np.testing.assert_allclose(actual_xarray_eofs_corr[0][0], desired_eofs_corr[0][0])
-    np.testing.assert_allclose(actual_xarray_eofs_corr[0][1], desired_eofs_corr[0][1])
-    np.testing.assert_allclose(actual_xarray_eofs_corr[1][0], desired_eofs_corr[1][0])
-    np.testing.assert_allclose(actual_xarray_eofs_corr[1][1], desired_eofs_corr[1][1])
-    np.testing.assert_allclose(actual_xarray_Xrec[0], desired_Xrec[0])
-    np.testing.assert_allclose(actual_xarray_Xrec[1], desired_Xrec[1])
-    np.testing.assert_allclose(actual_xarray_proj, desired_proj)
